third_party/WebKit/LayoutTests/fast/css3-text/css3-word-break/word-break-all-wrap-with-100percent-floats.html

<!DOCTYPE html>
<script src="../../../resources/ahem.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
p {
  width: 150px;
  border: 1px solid black;
  font: 10px Ahem;
}
span {
  float: left;
  width: 100%;
  height: 10px;
  background: papayawhip;
}
</style>
<div id="log"></div>
<div id="container">
  <div style="word-wrap: normal; word-break: normal">
    <p><span></span>Video source</p>
  </div>
  <div style="word-wrap: break-word">
    <p><span></span>Video source</p>
  </div>
  <div style="word-break: break-all">
    <p><span></span>Video source</p>
  </div>
</div>

<script>
Array.prototype.forEach.call(container.children, function (e) {
  test(function () {
    assert_equals(e.offsetHeight, 22, "height");
  }, e.style.cssText);
});
</script>
